+++ app/Http/Controllers/Settings/PermissionController.php
@@ -10,8 +10,21 @@ use Yajra\DataTables\DataTables;
 
 class PermissionController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = \Auth::guard('web')->user();
+            return $next($request);
+        });
+    }
+
     public function index(Request $request)
     {
+        #check jika user login || permisionya sesuai
+        if (is_null($this->user) || !$this->user->can('permission.view')) {
+            abort(403, 'Sorry! You are Unauthorized to see this page!');
+        }
+
         if ($request->ajax()) {
             $query = Permission::latest()->get();
 
@@ -35,6 +48,11 @@ class PermissionController extends Controller
 
     public function store(PermissionRequest $request)
     {
+        #check jika user login || permisionya sesuai
+        if (is_null($this->user) || !$this->user->can('permission.store')) {
+            abort(403, 'Sorry! You are Unauthorized to see this page!');
+        }
+
         $request->merge([
             'guard_name' => 'web'
         ]);
@@ -45,6 +63,11 @@ class PermissionController extends Controller
 
     public function destroy($id)
     {
+        #check jika user login || permisionya sesuai
+        if (is_null($this->user) || !$this->user->can('permission.delete')) {
+            abort(403, 'Sorry! You are Unauthorized to see this page!');
+        }
+
         $query = Permission::findOrFail($id);
         $query->delete();
 
@@ -56,6 +79,11 @@ class PermissionController extends Controller
 
     public function show($id)
     {
+        #check jika user login || permisionya sesuai
+        if (is_null($this->user) || !$this->user->can('permission.view')) {
+            abort(403, 'Sorry! You are Unauthorized to see this page!');
+        }
+
         $query = Permission::findOrFail($id);
 
         return response()->json([
@@ -67,6 +95,11 @@ class PermissionController extends Controller
 
     public function edit($id)
     {
+        #check jika user login || permisionya sesuai
+        if (is_null($this->user) || !$this->user->can('permission.edit')) {
+            abort(403, 'Sorry! You are Unauthorized to see this page!');
+        }
+
         $query = Permission::findOrFail($id);
 
         return response()->json([
@@ -78,6 +111,11 @@ class PermissionController extends Controller
 
     public function update(PermissionRequest $request)
     {
+        #check jika user login || permisionya sesuai
+        if (is_null($this->user) || !$this->user->can('permission.edit')) {
+            abort(403, 'Sorry! You are Unauthorized to see this page!');
+        }
+
         $query = Permission::findOrFail($request->permission_id);
         $query->update($request->all());
